bar refactor
