+++ docs/source/reference/progress_bar.rst
@@ -1,0 +1,7 @@
+rich.progress_bar
+========
+
+.. automodule:: rich.progress_bar
+    :members:
+
+
