+++ rich/bar.py
@@ -1,193 +1,85 @@
-import math
-from functools import lru_cache
-from time import monotonic
-from typing import Iterable, List, Optional
+from typing import Optional, Union
 
-from .color import Color, blend_rgb
-from .color_triplet import ColorTriplet
+from .color import Color
 from .console import Console, ConsoleOptions, RenderResult
 from .jupyter import JupyterMixin
 from .measure import Measurement
 from .segment import Segment
-from .style import Style, StyleType
+from .style import Style
 
-# Number of characters before 'pulse' animation repeats
-PULSE_SIZE = 20
+# There are left-aligned characters for 1/8 to 7/8, but
+# the right-aligned characters exist only for 1/8 and 4/8.
+BEGIN_BLOCK_ELEMENTS = ["█", "█", "█", "▐", "▐", "▐", "▕", "▕"]
+END_BLOCK_ELEMENTS = [" ", "▏", "▎", "▍", "▌", "▋", "▊", "▉"]
+FULL_BLOCK = "█"
 
 
 class Bar(JupyterMixin):
-    """Renders a (progress) bar.
+    """Renders a solid block bar.
 
     Args:
-        total (float, optional): Number of steps in the bar. Defaults to 100.
-        completed (float, optional): Number of steps completed. Defaults to 0.
+        size (float): Value for the end of the bar.
+        begin (float): Begin point (between 0 and size, inclusive).
+        end (float): End point (between 0 and size, inclusive).
         width (int, optional): Width of the bar, or ``None`` for maximum width. Defaults to None.
-        pulse (bool, optional): Enable pulse effect. Defaults to False.
-        style (StyleType, optional): Style for the bar background. Defaults to "bar.back".
-        complete_style (StyleType, optional): Style for the completed bar. Defaults to "bar.complete".
-        finished_style (StyleType, optional): Style for a finished bar. Defaults to "bar.done".
-        pulse_style (StyleType, optional): Style for pulsing bars. Defaults to "bar.pulse".
-        animation_time (Optional[float], optional): Time in seconds to use for animation, or None to use system time.
+        color (Union[Color, str], optional): Color of the bar. Defaults to "default".
+        bgcolor (Union[Color, str], optional): Color of bar background. Defaults to "default".
     """
 
     def __init__(
         self,
-        total: float = 100,
-        completed: float = 0,
+        size: float,
+        begin: float,
+        end: float,
+        *,
         width: int = None,
-        pulse: bool = False,
-        style: StyleType = "bar.back",
-        complete_style: StyleType = "bar.complete",
-        finished_style: StyleType = "bar.finished",
-        pulse_style: StyleType = "bar.pulse",
-        animation_time: float = None,
+        color: Union[Color, str] = "default",
+        bgcolor: Union[Color, str] = "default",
     ):
-        self.total = total
-        self.completed = completed
+        self.size = size
+        self.begin = max(begin, 0)
+        self.end = min(end, size)
         self.width = width
-        self.pulse = pulse
-        self.style = style
-        self.complete_style = complete_style
-        self.finished_style = finished_style
-        self.pulse_style = pulse_style
-        self.animation_time = animation_time
-
-        self._pulse_segments: Optional[List[Segment]] = None
+        self.style = Style(color=color, bgcolor=bgcolor)
 
     def __repr__(self) -> str:
-        return f"<Bar {self.completed!r} of {self.total!r}>"
-
-    @property
-    def percentage_completed(self) -> float:
-        """Calculate percentage complete."""
-        completed = (self.completed / self.total) * 100.0
-        completed = min(100, max(0.0, completed))
-        return completed
-
-    @lru_cache(maxsize=16)
-    def _get_pulse_segments(
-        self,
-        fore_style: Style,
-        back_style: Style,
-        color_system: str,
-        ascii: bool = False,
-    ) -> List[Segment]:
-        """Get a list of segments to render a pulse animation.
-
-        Returns:
-            List[Segment]: A list of segments, one segment per character.
-        """
-        bar = "-" if ascii else "━"
-        segments: List[Segment] = []
-
-        if color_system != "truecolor":
-            segments += [Segment(bar, fore_style)] * (PULSE_SIZE // 2)
-            segments += [Segment(bar if color_system else " ", back_style)] * (
-                PULSE_SIZE - (PULSE_SIZE // 2)
-            )
-            return segments
-
-        append = segments.append
-        fore_color = (
-            fore_style.color.get_truecolor()
-            if fore_style.color
-            else ColorTriplet(255, 0, 255)
-        )
-        back_color = (
-            back_style.color.get_truecolor()
-            if back_style.color
-            else ColorTriplet(0, 0, 0)
-        )
-        cos = math.cos
-        pi = math.pi
-        _Segment = Segment
-        _Style = Style
-        from_triplet = Color.from_triplet
-
-        for index in range(PULSE_SIZE):
-            position = index / PULSE_SIZE
-            fade = 0.5 + cos((position * pi * 2)) / 2.0
-            color = blend_rgb(fore_color, back_color, cross_fade=fade)
-            append(_Segment(bar, _Style(color=from_triplet(color))))
-        return segments
-
-    def update(self, completed: float, total: float = None) -> None:
-        """Update progress with new values.
-
-        Args:
-            completed (float): Number of steps completed.
-            total (float, optional): Total number of steps, or ``None`` to not change. Defaults to None.
-        """
-        self.completed = completed
-        self.total = total if total is not None else self.total
-
-    def _render_pulse(
-        self, console: Console, width: int, ascii: bool = False
-    ) -> Iterable[Segment]:
-        """Renders the pulse animation.
-
-        Args:
-            console (Console): Console instance.
-            width (int): Width in characters of pulse animation.
-
-        Returns:
-            RenderResult: [description]
-
-        Yields:
-            Iterator[Segment]: Segments to render pulse
-        """
-        fore_style = console.get_style(self.pulse_style, default="white")
-        back_style = console.get_style(self.style, default="black")
-
-        pulse_segments = self._get_pulse_segments(
-            fore_style, back_style, console.color_system, ascii=ascii
-        )
-        segment_count = len(pulse_segments)
-        current_time = (
-            monotonic() if self.animation_time is None else self.animation_time
-        )
-        segments = pulse_segments * (int(width / segment_count) + 2)
-        offset = int(-current_time * 15) % segment_count
-        segments = segments[offset : offset + width]
-        yield from segments
+        return f"Bar({self.size}, {self.begin}, {self.end})"
 
     def __rich_console__(
         self, console: Console, options: ConsoleOptions
     ) -> RenderResult:
 
         width = min(self.width or options.max_width, options.max_width)
-        ascii = options.legacy_windows or options.ascii_only
-        if self.pulse:
-            yield from self._render_pulse(console, width, ascii=ascii)
+
+        if self.begin >= self.end:
+            yield Segment(" " * width, self.style)
+            yield Segment.line()
             return
 
-        completed = min(self.total, max(0, self.completed))
+        prefix_complete_eights = int(width * 8 * self.begin / self.size)
+        prefix_bar_count = prefix_complete_eights // 8
+        prefix_eights_count = prefix_complete_eights % 8
 
-        bar = "-" if ascii else "━"
-        half_bar_right = " " if ascii else "╸"
-        half_bar_left = " " if ascii else "╺"
-        complete_halves = (
-            int(width * 2 * completed / self.total) if self.total else width * 2
-        )
-        bar_count = complete_halves // 2
-        half_bar_count = complete_halves % 2
-        style = console.get_style(self.style)
-        complete_style = console.get_style(
-            self.complete_style if self.completed < self.total else self.finished_style
-        )
-        _Segment = Segment
-        if bar_count:
-            yield _Segment(bar * bar_count, complete_style)
-        if half_bar_count:
-            yield _Segment(half_bar_right * half_bar_count, complete_style)
+        body_complete_eights = int(width * 8 * self.end / self.size)
+        body_bar_count = body_complete_eights // 8
+        body_eights_count = body_complete_eights % 8
 
-        remaining_bars = width - bar_count - half_bar_count
-        if remaining_bars and console.color_system is not None:
-            if not half_bar_count and bar_count:
-                yield _Segment(half_bar_left, style)
-                remaining_bars -= 1
-            if remaining_bars:
-                yield _Segment(bar * remaining_bars, style)
+        # When start and end fall into the same cell, we ideally should render
+        # a symbol that's "center-aligned", but there is no good symbol in Unicode.
+        # In this case, we fall back to right-aligned block symbol for simplicity.
+
+        prefix = " " * prefix_bar_count
+        if prefix_eights_count:
+            prefix += BEGIN_BLOCK_ELEMENTS[prefix_eights_count]
+
+        body = FULL_BLOCK * body_bar_count
+        if body_eights_count:
+            body += END_BLOCK_ELEMENTS[body_eights_count]
+
+        suffix = " " * (width - len(body))
+
+        yield Segment(prefix + body[len(prefix) :] + suffix, self.style)
+        yield Segment.line()
 
     def __rich_measure__(self, console: Console, max_width: int) -> Measurement:
         return (
@@ -195,19 +87,3 @@ class Bar(JupyterMixin):
             if self.width is not None
             else Measurement(4, max_width)
         )
-
-
-if __name__ == "__main__":  # pragma: no cover
-    console = Console()
-    bar = Bar(width=50, total=100)
-
-    import time
-
-    console.show_cursor(False)
-    for n in range(0, 101, 1):
-        bar.update(n)
-        console.print(bar)
-        console.file.write("\r")
-        time.sleep(0.05)
-    console.show_cursor(True)
-    console.print()
